Revamp the navigating flow of the main deck pages
This commit removes the last of that shadow second-top-nav in the
/$lang/index page, and now those links are available in the Deck
Overview, including the quick-search badge-link in the top right.
We also change some icons around and just generally make sure this
main-screen navigation makes sense, feels comfortable, and explains
to early testers what the intention of the page is meant to be.