Modify review cards dom structure (less wonky)
Previously we had a wonky structure where there was a single element for
the card contents whose values kept changing, and then the buttons were
all different elements rendered with a loop so that the mutations and
data cache could remain on the page. But this is confusing and makes it
more complicated to do anything with the page, and anyway it will no
longer work once we decide to do anything else with the cards like
animating a slide-in/slide-out.