Refactor Navbar icons
This commit removes the switch statement and the "renderIcon" function.
Now the NavbarData item wants `Icon: LucideIcon` rather than `icon: string`.
This will mean the navbar component actually isn't loading any of these
icons; they're being called on by the page itself and are being passed
through and render directly as <Icon /> within the navbar loop.

Theoretically this will be better for code-splitting later, but in any
case it's just much simpler code and we no longer have to remember which
strings match to which icons.