Make the navbar work using the AppNav's approach
This change pulls the contextMenu data point out into its own key
on the loaderData object, separate from titleBar, so now we have three
keys, (including appNav). And we convert to the AppNav's approach which
doesn't just get the data (if present) from the most recent match, but
goes through all matches with .findLast and finds the most recent match
that has a titleBar key, or that has a contextMenu key, and renders the
appropriate data. This way if a specific page adds a specific title but
doesn't do anything at all to the contextMenu, the contextMenu will still
inherit from the most recent match where it is populated, and vice versa.