Use window.location.origin -- not reactive, but fine
This commit adjusts a couple places where we were either hardcoding the URL
'https://sunlo.app' or using the env variable VITE_BASE_URL. This env var
is now no longer used at all, so it is removed from the docs and .env.example
file. It was a bad idea anyway; it is confusing w/r/t BASE_URL (provided
automatically by Vite's import.meta object), and perhaps a bit misleading as
it has no effect on the server's behavior. So now we can use BASE_URL if/when
we need it, use the router's useLocation when we need a reactive value, and
use window.location.origin when we don't (as with signup redirect links etc.)